Refactor cipher.go to return int64 for better compatibility

+++ cipher.go
@@ -3,6 +3,6 @@ package main
 import "io"
 
 type Cipher interface {
-	Encrypt([]byte, io.Writer, io.Reader) (int, error)
-	Decrypt([]byte, io.Writer, io.Reader) (int, error)
+	Encrypt([]byte, io.Writer, io.Reader) (int64, error)
+	Decrypt([]byte, io.Writer, io.Reader) (int64, error)
 }
